+++ frontend/src/app/members/friend/page.tsx
@@ -1,41 +1,27 @@
 'use client';
 
-import { useState, useEffect, useMemo } from 'react';
+import { useState, useEffect } from 'react';
 import toast from 'react-hot-toast';
 
 import { getFriends, acceptFriend, rejectFriend, deleteFriend, addFriend } from '@/api/friend';
 import { FriendDto, FriendStatus, FriendStatusMap } from '@/types/friend';
-import { COLORS } from '@/constants/colors';
-
-// 각 아코디언 패널의 열림/닫힘 상태를 관리
-type AccordionState = {
-  [key in FriendStatus]: boolean;
-};
-
-const AVATAR_COLORS = [
-  '#FFB6B6', '#B6E2FF', '#B6FFB6', '#FFF5B6', '#D1B6FF', '#FFB6E2', '#B6FFD1'
-];
-
-function getInitials(name: string) {
-  if (!name) return '';
-  const parts = name.split(' ');
-  if (parts.length === 1) return parts[0][0];
-  return parts[0][0] + parts[1][0];
-}
-
-function getAvatarColor(idx: number) {
-  return AVATAR_COLORS[idx % AVATAR_COLORS.length];
-}
+import FriendsItem from '@/components/domain/friend/FriendsItem';
+import AccordionPanel from '@/components/domain/friend/FriendsAccordion';
 
 const FriendsList: React.FC = () => {
+  // 친구 상태별 목록 처리
   const [acceptedFriends, setAcceptedFriends] = useState<FriendDto[]>([]);
   const [sentFriends, setSentFriends] = useState<FriendDto[]>([]);
   const [receivedFriends, setReceivedFriends] = useState<FriendDto[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+
+  // 친구 추가 위한 검색어(이메일)
   const [searchTerm, setSearchTerm] = useState<string>('');
 
-  const [accordionState, setAccordionState] = useState<AccordionState>({
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  // 아코디언 패널의 열림/닫힘 상태 관리
+  const [accordionState, setAccordionState] = useState({
     [FriendStatus.ACCEPTED]: true,
     [FriendStatus.SENT]: false,
     [FriendStatus.RECEIVED]: false,
@@ -45,10 +31,14 @@ const FriendsList: React.FC = () => {
 
   const fetchFriendsData = async () => {
     try {
-      const acceptedRes = await getFriends(FriendStatus.ACCEPTED);
-      const sentRes = await getFriends(FriendStatus.SENT);
-      const receivedRes = await getFriends(FriendStatus.RECEIVED);
+      // 각 상태별 친구 목록 조회
+      const [acceptedRes, sentRes, receivedRes] = await Promise.all([
+        getFriends(FriendStatus.ACCEPTED),
+        getFriends(FriendStatus.SENT),
+        getFriends(FriendStatus.RECEIVED),
+      ]);
 
+      // 각 상태별 친구 목록 세팅
       setAcceptedFriends(acceptedRes.data ?? []);
       setSentFriends(sentRes.data ?? []);
       setReceivedFriends(receivedRes.data ?? []);
@@ -61,10 +51,12 @@ const FriendsList: React.FC = () => {
     }
   };
 
+  // 최초 조회
   useEffect(() => {
     fetchFriendsData();
   }, []);
 
+  // 아코디언 패널 토글
   const handleToggleAccordion = (status: FriendStatus) => {
     setAccordionState(prev => ({
       ...prev,
@@ -72,41 +64,46 @@ const FriendsList: React.FC = () => {
     }));
   };
 
+  // 친구 목록 새로고침
   const refreshFriendsList = () => {
     setIsLoading(true);
     fetchFriendsData();
   };
 
+  // 친구 요청 수락
   const handleAcceptFriend = async (friendId: number) => {
     try {
-      await acceptFriend(friendId);
-      toast.success("친구가 수락되었습니다.");
+      const data = await acceptFriend(friendId);
+      toast.success(data.message || "친구가 수락되었습니다.");
       refreshFriendsList();
     } catch (e) {
       toast.error(e instanceof Error ? e.message : '친구 수락에 실패했습니다.');
     }
   };
 
+  // 친구 요청 거절
   const handleRejectFriend = async (friendId: number) => {
     try {
-      await rejectFriend(friendId);
-      toast.success("친구가 거절되었습니다.");
+      const data = await rejectFriend(friendId);
+      toast.success(data.message || "친구가 거절되었습니다.");
       refreshFriendsList();
     } catch (e) {
       toast.error(e instanceof Error ? e.message : '친구 거절에 실패했습니다.');
     }
   };
 
+  // 친구 삭제
   const handleDeleteFriend = async (friendId: number) => {
     try {
-      await deleteFriend(friendId);
-      toast.success("친구가 삭제되었습니다.");
+      const data = await deleteFriend(friendId);
+      toast.success(data.message || "친구가 삭제되었습니다.");
       refreshFriendsList();
     } catch (e) {
       toast.error(e instanceof Error ? e.message : '친구 삭제에 실패했습니다.');
     }
   };
 
+  // 친구 추가
   const handleFriendAdd = async () => {
     if (!searchTerm) {
       toast.error("친구의 이메일을 입력해주세요.");
@@ -114,8 +111,11 @@ const FriendsList: React.FC = () => {
     }
 
     try {
-      await addFriend(searchTerm);
-      toast.success(`${searchTerm} 님에게 친구 요청을 보냈습니다.`);
+      // 친구 추가 요청
+      const data = await addFriend(searchTerm);
+      toast.success(data.message || `${searchTerm} 님에게 친구 요청을 보냈습니다.`);
+
+      // 검색어 초기화, 친구 목록 재조회
       setSearchTerm('');
       refreshFriendsList();
     } catch (e) {
@@ -123,88 +123,6 @@ const FriendsList: React.FC = () => {
       toast.error(errorMessage);
     }
   };
-
-  const renderFriendItem = (friend: FriendDto, status: FriendStatus, idx: number) => {
-    if (typeof friend.friendId !== 'number') return null;
-    const avatarColor = getAvatarColor(idx);
-    return (
-      <div
-        key={friend.friendId}
-        className="flex items-center p-3 bg-white rounded-xl border border-gray-200 mb-3 shadow-sm gap-4 min-w-0 transition-shadow"
-      >
-        {friend.friendProfileImageUrl ? (
-          <div
-            className="w-[54px] h-[54px] rounded-full border-2 shadow-[0_2px_8px_0_rgba(0,0,0,0.07)] bg-[#f3f3f3] flex items-center justify-center overflow-hidden"
-            style={{
-              borderColor: avatarColor,
-            }}
-          >
-            <img
-              src={friend.friendProfileImageUrl}
-              alt={`${friend.friendNickname} 프로필`}
-              className="w-full h-full object-cover rounded-full"
-              style={{
-                aspectRatio: "1 / 1",
-                width: "100%",
-                height: "100%",
-                objectFit: "cover",
-                display: "block",
-              }}
-            />
-          </div>
-        ) : (
-          <div
-            className={`w-[54px] h-[54px] rounded-full flex items-center justify-center font-bold text-lg text-white border-2 shadow-[0_2px_8px_0_rgba(0,0,0,0.07)] select-none`}
-            style={{
-              backgroundColor: avatarColor,
-            }}
-          >
-            {getInitials(friend.friendNickname || '')}
-          </div>
-        )}
-        <div className="flex-grow min-w-0">
-          <div className="flex items-center gap-2.5">
-            <h3 className="m-0 text-[17px] font-bold text-gray-700 truncate max-w-[160px]">
-              {friend.friendNickname}
-            </h3>
-          </div>
-          <p className="mt-1 text-[13px] text-gray-400 truncate max-w-[280px]">
-            자기소개자기소개자기소개자기소개자기소개자기소개자기소개자기소개
-          </p>
-        </div>
-        <div className="flex gap-2.5">
-          {status === FriendStatus.ACCEPTED && (
-            <button
-              onClick={() => handleDeleteFriend(friend.friendId!)}
-              className="px-4 py-1.5 rounded-full text-white font-semibold text-[14px] min-w-[60px] h-9 bg-red-300 shadow transition hover:bg-red-600"
-            >
-              삭제
-            </button>
-          )}
-          {status === FriendStatus.RECEIVED && (
-            <>
-              <button
-                onClick={() => handleAcceptFriend(friend.friendId!)}
-                className="px-4 py-1.5 rounded-full text-white font-semibold text-[14px] min-w-[60px] h-9 bg-green-400 shadow transition hover:bg-green-700"
-              >
-                수락
-              </button>
-              <button
-                onClick={() => handleRejectFriend(friend.friendId!)}
-                className="px-4 py-1.5 rounded-full text-white font-semibold text-[14px] min-w-[60px] h-9 bg-gray-400 shadow transition hover:bg-gray-500"
-              >
-                거절
-              </button>
-            </>
-          )}
-        </div>
-      </div>
-    );
-  };
-
-  const filteredAcceptedFriends = acceptedFriends;
-  const filteredSentFriends = sentFriends;
-  const filteredReceivedFriends = receivedFriends;
 
   if (isLoading) {
     return (
@@ -234,7 +152,7 @@ const FriendsList: React.FC = () => {
           </span>
         </div>
 
-        {/* 검색과 친구추가 버튼 (수정) */}
+        {/* 검색과 친구추가 버튼 */}
         <div className="flex justify-end gap-2 mb-4 items-center w-full max-w-[540px]">
           <input
             type="text"
@@ -254,12 +172,22 @@ const FriendsList: React.FC = () => {
         {/* 아코디언: 친구인 목록 */}
         <AccordionPanel
           title={<>{FriendStatusMap[FriendStatus.ACCEPTED]}</>}
-          count={filteredAcceptedFriends.length}
+          count={acceptedFriends.length}
           isOpen={accordionState[FriendStatus.ACCEPTED]}
           onToggle={() => handleToggleAccordion(FriendStatus.ACCEPTED)}
         >
-          {filteredAcceptedFriends.length > 0 ? (
-            filteredAcceptedFriends.map((f, idx) => renderFriendItem(f, FriendStatus.ACCEPTED, idx))
+          {acceptedFriends.length > 0 ? (
+            acceptedFriends.map((f, idx) => (
+              <FriendsItem
+                key={f.friendId}
+                friend={f}
+                status={FriendStatus.ACCEPTED}
+                idx={idx}
+                onDelete={handleDeleteFriend}
+                onAccept={handleAcceptFriend}
+                onReject={handleRejectFriend}
+              />
+            ))
           ) : (
             <p className="text-center text-gray-300 text-[14px] m-0 py-2.5">
               <span role="img" aria-label="no-friends">🫥</span> 친구가 없습니다.
@@ -270,12 +198,22 @@ const FriendsList: React.FC = () => {
         {/* 아코디언: 친구 요청 받은 목록 */}
         <AccordionPanel
           title={<>{FriendStatusMap[FriendStatus.RECEIVED]}</>}
-          count={filteredReceivedFriends.length}
+          count={receivedFriends.length}
           isOpen={accordionState[FriendStatus.RECEIVED]}
           onToggle={() => handleToggleAccordion(FriendStatus.RECEIVED)}
         >
-          {filteredReceivedFriends.length > 0 ? (
-            filteredReceivedFriends.map((f, idx) => renderFriendItem(f, FriendStatus.RECEIVED, idx))
+          {receivedFriends.length > 0 ? (
+            receivedFriends.map((f, idx) => (
+              <FriendsItem
+                key={f.friendId}
+                friend={f}
+                status={FriendStatus.RECEIVED}
+                idx={idx}
+                onDelete={handleDeleteFriend}
+                onAccept={handleAcceptFriend}
+                onReject={handleRejectFriend}
+              />
+            ))
           ) : (
             <p className="text-center text-gray-300 text-[14px] m-0 py-2.5">
               <span role="img" aria-label="no-requests">😶‍🌫️</span> 받은 친구 요청이 없습니다.
@@ -286,12 +224,22 @@ const FriendsList: React.FC = () => {
         {/* 아코디언: 친구 요청한 목록 */}
         <AccordionPanel
           title={<>{FriendStatusMap[FriendStatus.SENT]}</>}
-          count={filteredSentFriends.length}
+          count={sentFriends.length}
           isOpen={accordionState[FriendStatus.SENT]}
           onToggle={() => handleToggleAccordion(FriendStatus.SENT)}
         >
-          {filteredSentFriends.length > 0 ? (
-            filteredSentFriends.map((f, idx) => renderFriendItem(f, FriendStatus.SENT, idx))
+          {sentFriends.length > 0 ? (
+            sentFriends.map((f, idx) => (
+              <FriendsItem
+                key={f.friendId}
+                friend={f}
+                status={FriendStatus.SENT}
+                idx={idx}
+                onDelete={handleDeleteFriend}
+                onAccept={handleAcceptFriend}
+                onReject={handleRejectFriend}
+              />
+            ))
           ) : (
             <p className="text-center text-gray-300 text-[14px] m-0 py-2.5">
               <span role="img" aria-label="no-sent">🫠</span> 보낸 친구 요청이 없습니다.
@@ -299,41 +247,6 @@ const FriendsList: React.FC = () => {
           )}
         </AccordionPanel>
       </div>
-    </div>
-  );
-};
-
-// 재사용 가능한 아코디언 패널 컴포넌트
-interface AccordionPanelProps {
-  title: React.ReactNode;
-  count: number;
-  isOpen: boolean;
-  onToggle: () => void;
-  children: React.ReactNode;
-}
-
-const AccordionPanel = ({ title, count, isOpen, onToggle, children }: AccordionPanelProps) => {
-  return (
-    <div className="mb-4 border border-gray-200 rounded-xl overflow-hidden bg-white shadow-sm w-full max-w-[540px] transition-shadow">
-      <div
-        onClick={onToggle}
-        className={`flex justify-between items-center px-4 py-2.5 cursor-pointer font-bold text-gray-700 text-[1.05rem] select-none transition-colors min-h-[36px] h-9 ${
-          isOpen ? 'bg-gray-50 border-b border-gray-200' : 'bg-white'
-        }`}
-      >
-        <span>
-          {title}{' '}
-          <span className="text-gray-400 font-semibold text-[14px]">({count})</span>
-        </span>
-        <span
-          className={`text-[1.15em] font-bold ml-2 transition-colors ${
-            isOpen ? 'text-blue-500' : 'text-gray-300'
-          }`}
-        >
-          {isOpen ? '▲' : '▼'}
-        </span>
-      </div>
-      {isOpen && <div className="px-4 pt-3 pb-2 bg-white">{children}</div>}
     </div>
   );
 };
